requirements.txt
altgraph==0.17.4
colorama==0.4.6
ffmpeg-python==0.2.0
future==1.0.0
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pywin32-ctypes==0.2.3
setuptools==80.9.0
tqdm==4.67.1
